Move modifyChangesetDiff to it's own folder away from lists to prep for changed items. Move currentChangesetId there too.
